Changed line-cap to "butt" and reduced sidewalk width on tertiary z13.
The line-cap for most roads was previously "round".  An extra-wide
casing is used to indicate "sidewalk" on secondary and tertiary roads,
and because all roads are drawn on top of all casings, the round end of
an unclassified road was drawn over the casing of a tertiary road it
joined.  Changing line-cap to "butt" resolves this in most cases,
although there is a bit of an artifact when a lesser road joins a
more major one as the more major one goes around the corner.

"sidewalk-width-z13" (which applies to both tertiaries and secondaries)
was reduced from 2 to 1.5, so that tertiaries don't have the sidewalk
colour dominating the road.  "sidewalk-width-z12" was left at 2, as
that only applies to secondaries, and the colour difference between
road colour and sidewalk (casing) colour there is less pronounced, so
less of a problem.